Add userId and presId to table
Started this feature to grab userId and presId from Relay's DB, but no
time to finish. To complete, add two colums to presentation_hits table:
presId(int) and userId(int) and test. With these two in the table, APIs
can be updated to match on userId instead of username or presPath.